Added JUnit tests and moved storage
