+++ src/main/java/seedu/nurseschedules/parser/Parser.java
@@ -23,7 +23,7 @@ public class Parser {
      * @param text User input
      * @return First word of user input
      */
-    public static String getFirstWord(String text) {
+    public String getFirstWord(String text) {
         int index = text.indexOf(' ');
 
         if (index > -1) {
